Transferred files from the Core to DataAccessLayer
