+++ DIMS-Core/DIMS-Core.BusinessLayer/Converters/DirectionConverter.cs
@@ -24,7 +24,7 @@ namespace DIMS_Core.BusinessLayer.Converters
         {
             var directionName = reader.GetString();
 
-            using var context = new DimsCoreContext();
+            using var context = new DIMSCoreContext();
             var direction = context.Directions.FirstOrDefault(x => x.Name == directionName);
 
             return direction?.DirectionId ?? 0;
@@ -42,7 +42,7 @@ namespace DIMS_Core.BusinessLayer.Converters
         {
             var directionId = value;
 
-            using var context = new DimsCoreContext();
+            using var context = new DIMSCoreContext();
             var direction = context.Directions.FirstOrDefault(x => x.DirectionId == directionId);
 
             if (direction != null)
