+++ DIMS-Core/DIMS-Core.DataAccessLayer/DIMS-Core.DataAccessLayer.csproj
@@ -27,4 +27,8 @@
     <ProjectReference Include="..\DIMS-Core.Common\DIMS-Core.Common.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Models" />
+  </ItemGroup>
+
 </Project>
