+++ DIMS-Core/DIMS-Core.DataAccessLayer/Interfaces/IUnitOfWork.cs
@@ -1,6 +1,6 @@
 ﻿using System;
-using System.Threading.Tasks;
 using DIMS_Core.DataAccessLayer.Models;
+using Task = System.Threading.Tasks.Task;
 
 namespace DIMS_Core.DataAccessLayer.Interfaces
 {
