+++ DIMS-Core/DIMS-Core.DataAccessLayer/Models/DIMSCoreContext.cs
@@ -1,119 +1,238 @@
-﻿using Microsoft.EntityFrameworkCore;
+using System;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata;
+
+#nullable disable
 
 namespace DIMS_Core.DataAccessLayer.Models
 {
-    public class DimsCoreContext : DbContext
+    public partial class DIMSCoreContext : DbContext
     {
-        public DimsCoreContext()
+        public DIMSCoreContext()
         {
         }
 
-        public DimsCoreContext(DbContextOptions<DimsCoreContext> options)
+        public DIMSCoreContext(DbContextOptions<DIMSCoreContext> options)
             : base(options)
         {
         }
 
         public virtual DbSet<Direction> Directions { get; set; }
-
+        public virtual DbSet<Task> Tasks { get; set; }
+        public virtual DbSet<TaskState> TaskStates { get; set; }
+        public virtual DbSet<TaskTrack> TaskTracks { get; set; }
         public virtual DbSet<UserProfile> UserProfiles { get; set; }
-
+        public virtual DbSet<UserTask> UserTasks { get; set; }
+        public virtual DbSet<VTask> VTasks { get; set; }
         public virtual DbSet<VUserProfile> VUserProfiles { get; set; }
+        public virtual DbSet<VUserTrack> VUserTracks { get; set; }
+
+        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+        {
+            if (!optionsBuilder.IsConfigured)
+            {
+#warning To protect potentially sensitive information in your connection string, you should move it out of source code. You can avoid scaffolding the connection string by using the Name= syntax to read it from configuration - see https://go.microsoft.com/fwlink/?linkid=2131148. For more guidance on storing connection strings, see http://go.microsoft.com/fwlink/?LinkId=723263.
+                optionsBuilder.UseSqlServer("Server=(localdb)\\mssqllocaldb;Database=DIMSCore;");
+            }
+        }
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
-            modelBuilder.Entity<Direction>(entity =>
-                                           {
-                                               entity.Property(e => e.Description)
-                                                     .HasMaxLength(250);
+            modelBuilder.HasAnnotation("Relational:Collation", "SQL_Latin1_General_CP1_CI_AS");
 
-                                               entity.Property(e => e.Name)
-                                                     .IsRequired()
-                                                     .HasMaxLength(50);
-                                           });
+            modelBuilder.Entity<Direction>(entity =>
+            {
+                entity.Property(e => e.Description).HasMaxLength(250);
+
+                entity.Property(e => e.Name)
+                    .IsRequired()
+                    .HasMaxLength(50);
+            });
+
+            modelBuilder.Entity<Task>(entity =>
+            {
+                entity.Property(e => e.DeadlineDate).HasColumnType("datetime");
+
+                entity.Property(e => e.Description).IsRequired();
+
+                entity.Property(e => e.Name)
+                    .IsRequired()
+                    .HasMaxLength(50);
+
+                entity.Property(e => e.StateDate).HasColumnType("datetime");
+            });
+
+            modelBuilder.Entity<TaskState>(entity =>
+            {
+                entity.HasKey(e => e.StateId)
+                    .HasName("PK__TaskStat__C3BA3B3A36B6EDFE");
+
+                entity.Property(e => e.StateName)
+                    .IsRequired()
+                    .HasMaxLength(50);
+            });
+
+            modelBuilder.Entity<TaskTrack>(entity =>
+            {
+                entity.Property(e => e.TrackDate).HasColumnType("datetime");
+
+                entity.Property(e => e.TrackNote)
+                    .IsRequired()
+                    .HasMaxLength(50);
+            });
 
             modelBuilder.Entity<UserProfile>(entity =>
-                                             {
-                                                 entity.HasKey(e => e.UserId)
-                                                       .HasName("PK__UserProf__1788CC4C01393682");
+            {
+                entity.HasKey(e => e.UserId)
+                    .HasName("PK_UserProfiles_UserId");
 
-                                                 entity.Property(e => e.Address)
-                                                       .HasMaxLength(120);
+                entity.Property(e => e.Address).HasMaxLength(120);
 
-                                                 entity.Property(e => e.BirthDate)
-                                                       .HasColumnType("date");
+                entity.Property(e => e.BirthDate).HasColumnType("date");
 
-                                                 entity.Property(e => e.Education)
-                                                       .HasMaxLength(50);
+                entity.Property(e => e.Education).HasMaxLength(50);
 
-                                                 entity.Property(e => e.Email)
-                                                       .IsRequired()
-                                                       .HasMaxLength(50);
+                entity.Property(e => e.Email)
+                    .IsRequired()
+                    .HasMaxLength(50);
 
-                                                 entity.Property(e => e.FirstName)
-                                                       .IsRequired()
-                                                       .HasMaxLength(50);
+                entity.Property(e => e.FirstName)
+                    .IsRequired()
+                    .HasMaxLength(50);
 
-                                                 entity.Property(e => e.LastName)
-                                                       .IsRequired()
-                                                       .HasMaxLength(50);
+                entity.Property(e => e.LastName)
+                    .IsRequired()
+                    .HasMaxLength(50);
 
-                                                 entity.Property(e => e.MobilePhone)
-                                                       .IsRequired()
-                                                       .HasMaxLength(50);
+                entity.Property(e => e.MobilePhone)
+                    .IsRequired()
+                    .HasMaxLength(50);
 
-                                                 entity.Property(e => e.Skype)
-                                                       .HasMaxLength(50);
+                entity.Property(e => e.Skype).HasMaxLength(50);
 
-                                                 entity.Property(e => e.StartDate)
-                                                       .HasColumnType("date");
+                entity.Property(e => e.StartDate).HasColumnType("date");
 
-                                                 entity.HasOne(d => d.Direction)
-                                                       .WithMany(p => p.UserProfile)
-                                                       .HasForeignKey(d => d.DirectionId)
-                                                       .OnDelete(DeleteBehavior.ClientSetNull)
-                                                       .HasConstraintName("FK__UserProfi__Direc__38996AB5");
-                                             });
+                entity.HasOne(d => d.Direction)
+                    .WithMany(p => p.UserProfiles)
+                    .HasForeignKey(d => d.DirectionId)
+                    .OnDelete(DeleteBehavior.ClientSetNull);
+            });
+
+            modelBuilder.Entity<UserTask>(entity =>
+            {
+                entity.Property(e => e.UserTaskId).ValueGeneratedOnAdd();
+
+                entity.HasOne(d => d.State)
+                    .WithMany(p => p.UserTasks)
+                    .HasForeignKey(d => d.StateId)
+                    .OnDelete(DeleteBehavior.ClientSetNull)
+                    .HasConstraintName("FK__UserTasks__State__6477ECF3");
+
+                entity.HasOne(d => d.Task)
+                    .WithMany(p => p.UserTasks)
+                    .HasForeignKey(d => d.TaskId)
+                    .OnDelete(DeleteBehavior.ClientSetNull)
+                    .HasConstraintName("FK__UserTasks__TaskI__656C112C");
+
+                entity.HasOne(d => d.User)
+                    .WithMany(p => p.UserTasks)
+                    .HasForeignKey(d => d.UserId)
+                    .OnDelete(DeleteBehavior.ClientSetNull)
+                    .HasConstraintName("FK__UserTasks__UserI__66603565");
+
+                entity.HasOne(d => d.UserTaskNavigation)
+                    .WithOne(p => p.UserTask)
+                    .HasForeignKey<UserTask>(d => d.UserTaskId)
+                    .OnDelete(DeleteBehavior.ClientSetNull)
+                    .HasConstraintName("FK__UserTasks__UserT__6383C8BA");
+            });
+
+            modelBuilder.Entity<VTask>(entity =>
+            {
+                entity.HasNoKey();
+
+                entity.ToView("vTask");
+
+                entity.Property(e => e.DeadlineDate).HasColumnType("datetime");
+
+                entity.Property(e => e.Description)
+                    .IsRequired()
+                    .HasColumnType("ntext");
+
+                entity.Property(e => e.Name)
+                    .IsRequired()
+                    .HasMaxLength(50);
+
+                entity.Property(e => e.StateDate).HasColumnType("datetime");
+
+                entity.Property(e => e.TaskId).ValueGeneratedOnAdd();
+            });
 
             modelBuilder.Entity<VUserProfile>(entity =>
-                                              {
-                                                  entity.HasNoKey();
+            {
+                entity.HasNoKey();
 
-                                                  entity.ToView("vUserProfiles");
+                entity.ToView("vUserProfiles");
 
-                                                  entity.Property(e => e.Address)
-                                                        .HasMaxLength(120);
+                entity.Property(e => e.Address).HasMaxLength(120);
 
-                                                  entity.Property(e => e.Direction)
-                                                        .IsRequired()
-                                                        .HasMaxLength(50);
+                entity.Property(e => e.Direction)
+                    .IsRequired()
+                    .HasMaxLength(50);
 
-                                                  entity.Property(e => e.Education)
-                                                        .HasMaxLength(50);
+                entity.Property(e => e.Education).HasMaxLength(50);
 
-                                                  entity.Property(e => e.Email)
-                                                        .IsRequired()
-                                                        .HasMaxLength(50);
+                entity.Property(e => e.Email)
+                    .IsRequired()
+                    .HasMaxLength(50);
 
-                                                  entity.Property(e => e.FullName)
-                                                        .IsRequired()
-                                                        .HasMaxLength(101);
+                entity.Property(e => e.FullName)
+                    .IsRequired()
+                    .HasMaxLength(101);
 
-                                                  entity.Property(e => e.MobilePhone)
-                                                        .IsRequired()
-                                                        .HasMaxLength(50);
+                entity.Property(e => e.MobilePhone)
+                    .IsRequired()
+                    .HasMaxLength(50);
 
-                                                  entity.Property(e => e.Skype)
-                                                        .HasMaxLength(50);
+                entity.Property(e => e.Skype).HasMaxLength(50);
 
-                                                  entity.Property(e => e.StartDate)
-                                                        .HasColumnType("date");
-                                              });
+                entity.Property(e => e.StartDate).HasColumnType("date");
+            });
+
+            modelBuilder.Entity<VUserTrack>(entity =>
+            {
+                entity.HasNoKey();
+
+                entity.ToView("vUserTrack");
+
+                entity.Property(e => e.Address).HasMaxLength(120);
+
+                entity.Property(e => e.Direction)
+                    .IsRequired()
+                    .HasMaxLength(50);
+
+                entity.Property(e => e.Education).HasMaxLength(50);
+
+                entity.Property(e => e.Email)
+                    .IsRequired()
+                    .HasMaxLength(50);
+
+                entity.Property(e => e.FullName)
+                    .IsRequired()
+                    .HasMaxLength(101);
+
+                entity.Property(e => e.MobilePhone)
+                    .IsRequired()
+                    .HasMaxLength(50);
+
+                entity.Property(e => e.Skype).HasMaxLength(50);
+
+                entity.Property(e => e.StartDate).HasColumnType("date");
+            });
 
             OnModelCreatingPartial(modelBuilder);
         }
 
-        private void OnModelCreatingPartial(ModelBuilder modelBuilder)
-        {
-        }
+        partial void OnModelCreatingPartial(ModelBuilder modelBuilder);
     }
 }
