+++ DIMS-Core/DIMS-Core.DataAccessLayer/Models/Direction.cs
@@ -1,20 +1,21 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+
+#nullable disable
 
 namespace DIMS_Core.DataAccessLayer.Models
 {
-    public class Direction
+    public partial class Direction
     {
         public Direction()
         {
-            UserProfile = new HashSet<UserProfile>();
+            UserProfiles = new HashSet<UserProfile>();
         }
 
         public int DirectionId { get; set; }
-
         public string Name { get; set; }
-
         public string Description { get; set; }
 
-        public virtual ICollection<UserProfile> UserProfile { get; set; }
+        public virtual ICollection<UserProfile> UserProfiles { get; set; }
     }
 }
