+++ DIMS-Core/DIMS-Core.DataAccessLayer/Models/VTask.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 #nullable disable
 
-namespace DIMS_Core.Models
+namespace DIMS_Core.DataAccessLayer.Models
 {
     public partial class VTask
     {
