+++ DIMS-Core/DIMS-Core.DataAccessLayer/Repositories/DirectionRepository.cs
@@ -4,7 +4,7 @@ namespace DIMS_Core.DataAccessLayer.Repositories
 {
     public class DirectionRepository : Repository<Direction>
     {
-        public DirectionRepository(DimsCoreContext context) : base(context)
+        public DirectionRepository(DIMSCoreContext context) : base(context)
         {
         }
     }
