+++ DIMS-Core/DIMS-Core.DataAccessLayer/Repositories/UserProfileRepository.cs
@@ -4,7 +4,7 @@ namespace DIMS_Core.DataAccessLayer.Repositories
 {
     public class UserProfileRepository : Repository<UserProfile>
     {
-        public UserProfileRepository(DimsCoreContext context) : base(context)
+        public UserProfileRepository(DIMSCoreContext context) : base(context)
         {
         }
     }
