+++ DIMS-Core/DIMS-Core.Tests/Repositories/DirectionRepositoryTests.cs
@@ -4,6 +4,7 @@ using DIMS_Core.DataAccessLayer.Models;
 using DIMS_Core.Tests.Repositories.Fixtures;
 using Microsoft.EntityFrameworkCore;
 using Xunit;
+using Task = System.Threading.Tasks.Task;
 
 namespace DIMS_Core.Tests.Repositories
 {
